% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectParams.R
\name{collectParams}
\alias{collectParams}
\title{Collect Parameters}
\usage{
collectParams(fExpr, mc = TRUE)
}
\arguments{
\item{fExpr}{An expression or a function object.}

\item{mc}{Flag to detect data for Monte Carlo sampling.}
}
\value{
A list containing:
\item{x.names}{vector of names for the input parameters}
\item{x.mu}{vector of mean values for the input parameters}
\item{x.u}{vector of standard uncertainties for the input parameters}
\item{x.pdf}{vector of pfd names for the input parameters}
\item{x.df}{vector of degrees of freedom for the input parameters}
\item{fExpr}{same as input}
}
\description{
Collects from the global environment the list of values
necessary to execute \code{GUMCV} and \code{GUMS1/2} on model \code{fExpr}.
The names of the variables should be consistent with the parameters of \code{fExpr}.
For a variable named \code{'var'}, the mean value is seached in variable \code{var},
the standard uncertainty in \code{var.mu}, the pdf in \code{var.pdf}, and
the degrees of freedom in \code{var.df}.
}
\examples{
fExpr = function(V,T,a,c) T*V + V^2/(2*a*c) 

V = 130 /3.6 ; V.u = 4/3.6         ; V.pdf = "norm"
T = 1        ; T.u = 0.5 / 3^0.5   ; T.pdf = "unif"
a = 5        ; a.u = 0.5           ; a.pdf = "norm"        
c = 0.7      ; c.u = 0.2/6         ; c.pdf = "norm"
# Collect parameters in environment
params=collectParams(fExpr)
# GUM
G = with(params, 
         gumCV(fExpr=fExpr,x.mu=x.mu,x.u=x.u)
         )
}

