% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arcsine.R
\name{darcsine}
\alias{darcsine}
\alias{parcsine}
\alias{qarcsine}
\alias{rarcsine}
\title{The Arcsine Distribution}
\usage{
darcsine(x, a = 0, b = 1)

parcsine(q, a = 0, b = 1)

qarcsine(p, a = 0, b = 1)

rarcsine(n, a = 0, b = 1)
}
\arguments{
\item{\code{x,q}}{vector of quantiles}

\item{\code{p}}{vector of probabilities}

\item{\code{n}}{number of random draws}
}
\description{
Density, distribution function, quantile function and random generation for
the arcsine distribution on the interval from \code{a} to \code{b}.
}
\examples{
darcsine(seq(0,1,by=0.1))
parcsine(seq(-pi,pi,by=0.1), a=-pi, b=pi)
qarcsine(seq(0,1,by=0.1), a=-pi, b=pi)
rarcsine(10)
}

