% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ECIPlot.R
\name{ECIPlot}
\alias{ECIPlot}
\title{Plot of Empirical Cumulated Density function and p-percent Coverage Interval}
\usage{
ECIPlot(Y, p = 0.95, cex = 1)
}
\arguments{
\item{Y}{A vector.}

\item{p}{Coverage of confidence interval.}

\item{cex}{Graphical parameter.}
}
\value{
A plot.
}
\description{
Generates plot from sample.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
ECIPlot(S$Y)
}

