% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gumCV.R
\name{gumCV}
\alias{gumCV}
\title{Combination of Variances}
\usage{
gumCV(fExpr, x.mu, x.u, x.cor = diag(length(x.mu)), x.cov = NULL,
  budgetTable = TRUE, silent = FALSE)
}
\arguments{
\item{\code{fExpr}}{an expression or a function object}

\item{\code{x.mu}}{named vector of mean values
with names compatible with \code{fExpr}}

\item{\code{x.u}}{named vector of standard uncertainty values
(one of {x.u, x.cov} oblig)}

\item{\code{x.cor}}{named correlation matrix between model parameters}

\item{\code{x.cov}}{named variance/covariance matrix
between model parameters (one of {x.u, x.cov} oblig)}

\item{\code{budgetTable}}{whether the budget table is computed}

\item{\code{silent}}{whether gumCV executes without printout}
}
\value{
A list containing:
\item{y.mu}{mean value of model}
\item{y.u}{standard uncertainty of model}
\item{anova}{(vector) relative contributions of parameters to y variance}
\item{anovaCov}{global relative contribution of parameters covariance to y variance}
\item{budget}{(dataframe) uncertainty budget table, mostly to be printed}
}
\description{
Uncertainty propagation by combination of variances (GUM).
}
\references{
(GUM) Evaluation of measurement data – Guide to the expression of
uncertainty in measurement. JCGM 100:2008.
\url{http://www.bipm.org/utils/common/documents/jcgm/JCGM_100_2008_E.pdf}
}

