% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gumMC.R
\name{gumS1}
\alias{gumS1}
\alias{gumS2}
\title{Combination of Distributions by Monte Carlo}
\usage{
gumS1(fExpr, x.mu, x.u, x.pdf, x.df, x.cor = diag(length(x.mu)),
  x.cov = NULL, nrun = 1000, adapt = FALSE, ndig = 1, p = 0.95,
  delFrac = 1, stdev = TRUE, interval = TRUE, silent = FALSE,
  nrunMax = 1e+06)

gumS2(fExpr, x.mu, x.u, x.pdf, x.df, x.cor = diag(length(x.mu)),
  x.cov = NULL, nrun = 100, h1 = 30, ndig = 1, p = 0.95,
  silent = FALSE, nrunMax = 1e+06)
}
\arguments{
\item{\code{fExpr}}{an expression or a function object}

\item{\code{x.mu}}{named vector of mean values
with names compatible with \code{fExpr}}

\item{\code{x.u}}{named vector of standard uncertainty values
(one of {x.u, x.cov} oblig)}

\item{\code{x.pdf}}{named vector of pdf types (norm, unif, stud...)}

\item{\code{x.df}}{named vector of degrees of freedom for \code{x.pdf}}

\item{\code{x.cor}}{named correlation matrix between model parameters}

\item{\code{x.cov}}{named variance/covariance matrix
between model parameters (one of {x.u, x.cov} oblig)}

\item{\code{nrun}}{number of runs in each sample packet}

\item{\code{h1}}{number of packets in first step of \code{gumS2}}

\item{\code{adapt}}{whether use the sequential adaptive method}

\item{\code{p}}{coverage of confidence interval}

\item{\code{stdev}}{whether to converge standard deviation (\code{gumS1};
for \code{gumS2} convergence is based on the variance)}

\item{\code{interval}}{whether to converge confidence interval (\code{gumS1})}

\item{\code{ndig}}{number of significant figures to converge}

\item{\code{nrunMax}}{maximum number of runs allowed}

\item{\code{silent}}{whether function executes without printout}
}
\value{
A list containing:
\item{y.mu}{mean value of model}
\item{y.u}{standard uncertainty of model}
\item{y.low}{lower limit of confidence interval}
\item{y.high}{upper limit of confidence interval}
\item{p}{coverage of confidence interval (same as input)}
\item{X}{(matrix) sample of inputs used to converge statistics}
\item{Y}{(vector) sample of outputs corresponding to X}
}
\description{
Uncertainty propagation by Monte Carlo (GUM-Supp1 [1]).

Both \code{gumS1(adapt=TRUE)} and \code{gumS2} provide adaptative procedures.
\code{gumS1} implements the code proposed in GUM-Supp1 which checks the
convergence of the estimators by blocks.
\code{gumS2} implements the two-stage Stein procedure [2], which has
been shown to have a less erratic stopping criterion than \code{gumS1}.

By default, \code{gumS1} is non-adaptative and performs the user-specified
number of simulations.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
pairs(cbind(S$X,S$Y))
S=gumS2(fExpr,x.mu,x.u,x.pdf,x.df=NULL)
ECIPlot(S$Y)
}
\references{
[1] Evaluation of measurement data – Supplement 1 to
the "Guide to the expression of uncertainty in measurement" –
Propagation of distributions using a Monte Carlo method. \emph{JCGM} \strong{101}:2008.
\href{http://www.bipm.org/utils/common/documents/jcgm/JCGM_101_2008_E.pdf}{PDF}

[2] G. W\"ubbeler, P. M. Harris, M. G. Cox and C. Elster (2010)
A two-stage procedure for determining the number of trials
in the application of a Monte Carlo method for uncertainty evaluation.
\emph{Metrologia} \strong{47}:317.
\href{http://dx.doi.org/10.1088/0026-1394/47/3/023}{CrossRef}
}

