% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uncPrint.R
\name{uncPrint}
\alias{CIPrint}
\alias{CIPrint1}
\alias{uncPrint}
\title{Printing Functions}
\usage{
uncPrint(y, uy)

CIPrint(y, uy, p = 0.95, fac = 1.96)

CIPrint1(uy, y_low, y_high, p = 0.95)
}
\arguments{
\item{\code{y}}{mean value}

\item{\code{uy}}{standard uncertainty on \code{y}}

\item{\code{p}}{coverage of confidence interval}

\item{\code{fac}}{enlargment factor corresponding to \code{p}}

\item{\code{y_low,y_high}}{lower and upper limits of coverage interval}
}
\value{
A character string.
}
\description{
Print results with prescribed number of
digits.

\code{uncPrint()} is used to print the mean value and
(enlarged) standard uncertainty.
\code{CIPrint()} is used to print coverage intervals
from mean value, standard uncertainty and coverage factor.
\code{CIPrint1()} is used to print coverage intervals
from interval limits.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
uncPrint(S$y.mu,S$y.u)
#
CIPrint(S$y.mu,S$y.u)
#
CIPrint1(S$y.u,S$y.low,S$y.high)
}

