% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsicSA.R
\name{hsicSA}
\alias{hsicSA}
\title{Hilbert-Schmidt Independence Criterion Sensitivity Analysis}
\usage{
hsicSA(X, Y, budgetTable = TRUE, silent = FALSE)
}
\arguments{
\item{X}{A \code{M}x\code{N} matrix of \code{M} values for 
\code{N} input variables.}

\item{Y}{A vector of values corresponding to \code{X}.}

\item{budgetTable}{Flag to compute the budget table.}

\item{silent}{Flag to run without printout.}
}
\value{
A list containing:
\item{hs}{(vector) HSIC sensitivity indices}
\item{budget}{(dataframe) VG budget table, mostly to be printed}
}
\description{
Uses Monte Carlo sample to generate Hilbert-Schmidt Independence Criterion
based SA indices
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S = gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
hsicSA(S$X,S$Y,silent=TRUE)
}
\references{
Da Veiga S. (2015), 
Global sensitivity analysis with dependence measures, 
\emph{J. Stat. Comp. Sim.}, \strong{85}:1283-1305.
}
