% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SAPlot.R
\name{SAPlot}
\alias{SAPlot}
\title{Sensitivity Analysis Plot}
\usage{
SAPlot(X, cex = 1)
}
\arguments{
\item{X}{A \code{M}x\code{N} matrix of \code{M} values for \code{N} variables.}

\item{cex}{Graphical parameter.}
}
\value{
A plot.
}
\description{
2D scatterplots of the inputs/outputs sample, histograms and correlation coefficients.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
SAPlot(cbind(S$X,S$Y))
}

