% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolSA.R
\name{sobolSA}
\alias{sobolSA}
\title{Sobol Sensitivity Indices}
\usage{
sobolSA(fExpr, x.mu, x.u, x.pdf, x.df, M = 10000, nboot = 0, graph = TRUE,
  cex = 1)
}
\arguments{
\item{fExpr}{An expression or a function object.}

\item{x.mu}{Named vector of mean values 
with names compatible with \code{fExpr}.}

\item{x.u}{Named vector of standard uncertainty values 
(one of \{\code{x.u}, \code{x.cov}\} mandatory).}

\item{x.pdf}{Named vector of pdf types (see \code{\link{PDFs}}).}

\item{x.df}{Named vector of degrees of freedom for \code{x.pdf}.}

\item{M}{Size of Sobol samples.}

\item{nboot}{Size of bootstrap to calculate uncertainty on Sobol indices.}

\item{graph}{Flag to plot bargraphs of Sobol indices.}

\item{cex}{A graphical parameter.}
}
\value{
See \code{\link[sensitivity]{soboljansen}}.
}
\description{
Compute and plots Sobol first- and total-order sensitivity indices for model \code{fExpr}. 
Only for independeninput parameters. This is a wrapper for the \code{soboljansen}
function of package \code{sensitivity}.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
sobolSA(fExpr,x.mu,x.u,x.pdf)
}

