% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSample.R
\name{xSample}
\alias{PDFs}
\alias{xSample}
\title{Generate Sample}
\usage{
xSample(M = 1000, x.mu, x.u, x.pdf, x.df, x.cor = diag(length(x.mu)),
  x.cov = NULL, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{M}{Sample size.}

\item{x.mu}{Named vector of mean values 
with names compatible with \code{fExpr}.}

\item{x.u}{Named vector of standard uncertainty values 
(one of \{\code{x.u}, \code{x.cov}\} mandatory).}

\item{x.pdf}{Named vector of pdf types (see below).}

\item{x.df}{Named vector of degrees of freedom for \code{x.pdf}.}

\item{x.cor}{Named correlation matrix between model parameters.}

\item{x.cov}{Named variance/covariance matrix between model parameters 
(one of \{\code{x.u}, \code{x.cov}\} mandatory).}

\item{tol}{Numeric tolerance level to check positive-definiteness of 
\code{x.cor} or \code{x.cov}.}
}
\value{
\code{X} A \code{M}*\code{N} matrix of \code{M} values for \code{N} 
         variables.
}
\description{
Generate Monte Carlo sample from prescribed PDF. 
All PDFs are parameterized by their mean and standard deviation.
}
\details{
Available distributions:
\itemize{
\item{'delta'} {Dirac delta distribution for constants; args = x.mu}
\item{'norm'} {Normal; args = x.mu, x.u}
\item{'tnorm'} {Truncated normal (positive values); args = x.mu, x.u}
\item{'lnorm'} {Lognormal; args = x.mu, x.u}
\item{'stud'} {Student's T; args = x.mu, x.u, x.df}
\item{'unif'} {Uniform; args = x.mu, x.u}
\item{'triangle'} {Symmetric triangular; args = x.mu, x.u}
\item{'arcsine'} {Arcsine derivative; args = x.mu, x.u}
\item{'pois'} {Poisson; args = x.mu}
}

Correlation between variables is described by matrices \code{x.cov}
or \code{x.cor}, and enforced by Gaussian copula.
}
\examples{
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
X=xSample(M=1000,x.mu,x.u,x.pdf)
SAPlot(X)
}

