% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gumCV.R
\name{gumCV}
\alias{gumCV}
\title{Combination of Variances}
\usage{
gumCV(fExpr, x.mu, x.u, x.cor = diag(length(x.mu)), x.cov = NULL,
  budgetTable = TRUE, silent = FALSE, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{fExpr}{An expression or a function object. For an expression,
all variables will be considered as uncertain. For a
function the list of uncertain variables is derived
from the list of arguments.}

\item{x.mu}{Named vector of mean values
with names compatible with \code{fExpr}.}

\item{x.u}{Named vector of standard uncertainty values
(one of \{\code{x.u}, \code{x.cov}\} mandatory).}

\item{x.cor}{Named correlation matrix between model parameters}

\item{x.cov}{Named variance/covariance matrix
between model parameters
(one of \{\code{x.u}, \code{x.cov}\} mandatory).}

\item{budgetTable}{Flag to controle cmputation of the budget table.}

\item{silent}{Flag to execute \code{gumCV} without printout.}

\item{tol}{Numeric tolerance level to check positive-definiteness of
                   \code{x.cor} or \code{x.cov}.}
}
\value{
A list containing:
\item{y.mu}{mean value of model}
\item{y.u}{standard uncertainty of model}
\item{anova}{(vector) relative contributions of parameters to y variance}
\item{anovaCov}{global relative contribution of parameters covariance to \code{Y} variance}
\item{budget}{(dataframe) uncertainty budget table, mostly to be printed}
}
\description{
Uncertainty propagation by combination of variances (GUM).
}
\references{
(GUM) Evaluation of measurement data - Guide to the expression of
uncertainty in measurement. JCGM 100:2008.
\url{http://www.bipm.org/utils/common/documents/jcgm/JCGM_100_2008_E.pdf}
}

