% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumMC.R
\name{gumS1}
\alias{gumS1}
\alias{gumS2}
\title{Combination of Distributions by Monte Carlo}
\usage{
gumS1(fExpr, x.mu, x.u, x.pdf, x.df, x.cor = diag(length(x.mu)),
  x.cov = NULL, nrun = 1000, adapt = FALSE, ndig = 1, p = 0.95,
  delFrac = 1, stdev = TRUE, interval = TRUE, silent = FALSE,
  nrunMax = 1e+06)

gumS2(fExpr, x.mu, x.u, x.pdf, x.df, x.cor = diag(length(x.mu)),
  x.cov = NULL, nrun = 100, h1 = 30, ndig = 1, p = 0.95,
  silent = FALSE, nrunMax = 1e+06)
}
\arguments{
\item{fExpr}{An expression or a function object.}

\item{x.mu}{Named vector of mean values 
with names compatible with \code{fExpr}.}

\item{x.u}{Named vector of standard uncertainty values 
(one of \{\code{x.u}, \code{x.cov}\} mandatory).}

\item{x.pdf}{Named vector of pdf types (see \code{\link{PDFs}}).}

\item{x.df}{Named vector of degrees of freedom for \code{x.pdf}.}

\item{x.cor}{Named correlation matrix between model parameters.}

\item{x.cov}{Named variance/covariance matrix 
between model parameters (one of \{\code{x.u}, 
\code{x.cov}\} mandatory).}

\item{nrun}{Number of runs in each sample packet.}

\item{adapt}{Flag to use the sequential adaptive method.}

\item{ndig}{Number of significant figures to converge.}

\item{p}{Coverage of confidence interval.}

\item{delFrac}{Multiplicative factor on  numerical tolerance. 
               For compatibility with examples in GUM-Supp1 [1].
#'}

\item{stdev}{Flag for adaptative method to converge standard 
deviation (\code{gumS1}; for \code{gumS2} convergence 
is based on the variance).}

\item{interval}{Flag to converge confidence interval (\code{gumS1}).}

\item{silent}{Flag to run without printout.}

\item{nrunMax}{Maximum number of runs allowed.}

\item{h1}{Number of packets in first step of \code{gumS2}.}
}
\value{
A list containing:
\item{y.mu}{mean value of model}
\item{y.u}{standard uncertainty of model}
\item{y.low}{lower limit of confidence interval}
\item{y.high}{upper limit of confidence interval}
\item{p}{coverage of confidence interval (same as input)}
\item{X}{(matrix) sample of inputs used to converge statistics}
\item{Y}{(vector) sample of outputs corresponding to X}
}
\description{
Uncertainty propagation by Monte Carlo (GUM-Supp1 [1]).

Both \code{gumS1(adapt=TRUE)} and \code{gumS2} provide adaptative procedures. 
\code{gumS1} implements the code proposed in GUM-Supp1 which checks the 
convergence of the estimators by blocks. 
\code{gumS2} implements the two-stage Stein procedure [2], which has 
been shown to have a less erratic stopping criterion than \code{gumS1}.

By default, \code{gumS1} is non-adaptative and performs the user-specified
number of simulations.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
pairs(cbind(S$X,S$Y))
S=gumS2(fExpr,x.mu,x.u,x.pdf,x.df=NULL)
ECIPlot(S$Y)
}
\references{
[1] Evaluation of measurement data - Supplement 1 to 
the "Guide to the expression of uncertainty in measurement" -  
Propagation of distributions using a Monte Carlo method. \emph{JCGM} \strong{101}:2008.
\href{http://www.bipm.org/utils/common/documents/jcgm/JCGM_101_2008_E.pdf}{PDF}

[2] G. W\"ubbeler, P. M. Harris, M. G. Cox and C. Elster (2010) 
A two-stage procedure for determining the number of trials 
in the application of a Monte Carlo method for uncertainty evaluation.
\emph{Metrologia} \strong{47}:317. 
\href{http://dx.doi.org/10.1088/0026-1394/47/3/023}{CrossRef}
}
