% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncPrint.R
\name{uncPrint}
\alias{CIPrint}
\alias{CIPrint1}
\alias{UncPrint}
\alias{uncPrint}
\title{Printing Functions}
\usage{
uncPrint(y, uy, numDig = 2)

UncPrint(y, uy, fac = 2)

CIPrint(y, uy, p = 0.95, fac = 1.96)

CIPrint1(uy, y_low, y_high, p = 0.95)
}
\arguments{
\item{y}{Mean value.}

\item{uy}{Standard uncertainty on \code{Y}. Should be strictly positive.}

\item{fac}{Enlargment factor corresponding to \code{p}.}

\item{p}{Coverage of confidence interval.}

\item{y_low, y_high}{Lower and upper limits of coverage interval.}
}
\value{
A character string.
}
\description{
Print results with prescribed number of
digits following GUM recommendations.

\code{uncPrint()} is used to print the mean value and
standard uncertainty.
\code{UncPrint()} is used to print the mean value and
enlarged uncertainty.
\code{CIPrint()} is used to print coverage intervals 
from mean value, standard uncertainty and coverage factor.
\code{CIPrint1()} is used to print coverage intervals 
from interval limits.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
uncPrint(S$y.mu,S$y.u)
#
UncPrint(S$y.mu,2*S$y.u,fac=2)
#
CIPrint(S$y.mu,S$y.u)
#
CIPrint1(S$y.u,S$y.low,S$y.high)
}

