% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sobolSA.R
\name{sobolSA}
\alias{sobolSA}
\title{Sobol Sensitivity Indices}
\usage{
sobolSA(fExpr, x.mu, x.u, x.pdf, x.df, M = 10000, nboot = 0, graph = TRUE,
  cex = 1)
}
\arguments{
\item{\code{fExpr}}{an expression or a function object}

\item{\code{x.mu}}{named vector of mean values
with names compatible with \code{fExpr}}

\item{\code{x.u}}{named vector of standard uncertainty values
(one of {x.u, x.cov} oblig)}

\item{\code{x.pdf}}{named vector of pdf types (norm, unif, stud...)}

\item{\code{x.df}}{named vector of degrees of freedom for \code{x.pdf}}

\item{\code{M}}{size of Sobol samples}

\item{\code{nboot}}{size of bootstrap to calculate uncertainty on Sobol indices}

\item{\code{graph}}{(logical) if TRUE (default), plot bargraphs of Sobol indices}

\item{\code{cex}}{a graphical parameter}
}
\value{
See \code{\link[sensitivity]{soboljansen}}.
}
\description{
Compute and plots Sobol first- and total-order sensitivity indices for model \code{fExpr}.
Only for independeninput parameters. This is a wrapper for the \code{soboljansen}
function of package \code{sensitivity}.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
sobolSA(fExpr,x.mu,x.u,x.pdf)
}

