% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vgSA.R
\name{vgSA}
\alias{vgSA}
\title{Variance Gradient Sensitivity Analysis}
\usage{
vgSA(fExpr, x.mu, x.u, X, Y, budgetTable = TRUE, silent = FALSE)
}
\arguments{
\item{\code{fExpr}}{an expression or a function object}

\item{\code{x.mu}}{named vector of mean values
with names compatible with \code{fExpr}}

\item{\code{x.u}}{named vector of standard uncertainty values}

\item{\code{X}}{a MxN matrix of M values for N input variables}

\item{\code{Y}}{a vector of values corresponding to \code{X}}

\item{\code{budgetTable}}{whether the budget table is computed}

\item{\code{silent}}{whether gumCV executes without printout}
}
\value{
A list containing:
\item{vg}{(vector) variance gradients}
\item{budget}{(dataframe) VG budget table, mostly to be printed}
}
\description{
Uses Monte Carlo sample to generate Variance Gradients estimates.
Only for independent input parameters.
}
\examples{
fExpr = expression(x1+x2)
x.mu = c(1,1); names(x.mu)=c('x1','x2')
x.u = c(0.1,0.1); names(x.u)=c('x1','x2')
x.pdf = c('unif','triangle'); names(x.pdf)=c('x1','x2')
S=gumS1(fExpr,x.mu,x.u,x.pdf,x.df=NULL,nrunMax=1000)
vgSA(fExpr,x.mu,x.u,S$X,S$Y)
}
\references{
Mark Campanelli, Raghu Kacker and R\"udiger Kessel (2013)
Variance gradients and uncertainty budgets for nonlinear measurement
functions with independent inputs. \emph{Meas. Sci. Technol.} \strong{24}:25002.
}

